fix: extract build-id into seperate package to fix publishing
Also makes it easier to replace during bundling.
