+++ build-id/README.md
@@ -1,0 +1,3 @@
+# Fresh build id
+
+Don't use this package directly. It's considered internal for Fresh.
