+++ build-id/deno.json
@@ -1,0 +1,11 @@
+{
+  "name": "@fresh/build-id",
+  "version": "1.0.0",
+  "license": "MIT",
+  "exports": {
+    ".": "./mod.ts"
+  },
+  "imports": {
+    "@std/encoding": "jsr:@std/encoding@^1.0.10"
+  }
+}
