+++ deno.lock
@@ -90,6 +90,8 @@
     "npm:tailwindcss@^4.1.10": "4.1.11",
     "npm:ts-morph@26": "26.0.0",
     "npm:vite-plugin-inspect@^11.3.2": "11.3.2_vite@7.0.6__@types+node@22.15.15__picomatch@4.0.3_@types+node@22.15.15",
+    "npm:vite@*": "7.0.6_@types+node@22.15.15_picomatch@4.0.3",
+    "npm:vite@7.0.6": "7.0.6_@types+node@22.15.15_picomatch@4.0.3",
     "npm:vite@^7.0.6": "7.0.6_@types+node@22.15.15_picomatch@4.0.3"
   },
   "jsr": {
@@ -2855,6 +2857,7 @@
       "jsr:@astral/astral@~0.5.3",
       "jsr:@deno/doc@0.172",
       "jsr:@deno/esbuild-plugin@^1.1.5",
+      "jsr:@fresh/build-id@1",
       "jsr:@fresh/core@^2.0.0-alpha.37",
       "jsr:@marvinh-test/fresh-island@^0.0.1",
       "jsr:@std/async@^1.0.13",
@@ -2892,6 +2895,11 @@
       "npm:ts-morph@26"
     ],
     "members": {
+      "build-id": {
+        "dependencies": [
+          "jsr:@std/encoding@^1.0.10"
+        ]
+      },
       "plugin-tailwindcss-v3": {
         "dependencies": [
           "npm:autoprefixer@^10.4.21",
