+++ src/build_cache.ts
@@ -4,7 +4,7 @@ import { fsItemsToCommands, type FsRouteFile } from "./fs_routes.ts";
 import type { ServerIslandRegistry } from "./context.ts";
 import type { AnyComponent } from "preact";
 import { UniqueNamer } from "./utils.ts";
-import { setBuildId } from "fresh/build-id";
+import { setBuildId } from "@fresh/build-id";
 
 export interface FileSnapshot {
   name: string;
