+++ src/dev/builder.ts
@@ -17,7 +17,7 @@ import {
   type FsRoute,
   MemoryBuildCache,
 } from "./dev_build_cache.ts";
-import { BUILD_ID } from "../runtime/build_id.ts";
+import { BUILD_ID } from "@fresh/build-id";
 import { updateCheck } from "./update_check.ts";
 import { devErrorOverlay } from "./middlewares/error_overlay/middleware.tsx";
 import { automaticWorkspaceFolders } from "./middlewares/automatic_workspace_folders.ts";
